1. LangChain Introduction/requirements.txt
langchain
langchain_community
langchain_openai
langchain_huggingface
openai
huggingface_hub
python-dotenv
streamlit
ipykernel
